Api_liliang/requirement.txt
certifi==2019.3.9
chardet==3.0.4
ddt==1.2.1
et-xmlfile==1.0.1
idna==2.8
jdcal==1.4.1
openpyxl==2.6.2
PyMySQL==0.9.3
python-hessian==1.1.0
requests==2.22.0
six==1.12.0
suds-jurko==0.6
urllib3==1.25.2
